Moved log.py to logging
Also renamed logging/logging.py to logging/rich.py
