+++ tests/cmds/test_send_message.py
@@ -25,7 +25,7 @@ class TestSendMessageCommand(unittest.TestCase):
         return aprsd_config.Config(config)
 
     @mock.patch("aprsd.config.parse_config")
-    @mock.patch("aprsd.log.setup_logging")
+    @mock.patch("aprsd.logging.log.setup_logging")
     def test_no_login(self, mock_logging, mock_parse_config):
         """Make sure we get an error if there is no login and config."""
 
@@ -43,7 +43,7 @@ class TestSendMessageCommand(unittest.TestCase):
         assert "Must set --aprs_login or APRS_LOGIN" in result.output
 
     @mock.patch("aprsd.config.parse_config")
-    @mock.patch("aprsd.log.setup_logging")
+    @mock.patch("aprsd.logging.log.setup_logging")
     def test_no_password(self, mock_logging, mock_parse_config):
         """Make sure we get an error if there is no password and config."""
 
@@ -58,7 +58,7 @@ class TestSendMessageCommand(unittest.TestCase):
         assert "Must set --aprs-password or APRS_PASSWORD" in result.output
 
     @mock.patch("aprsd.config.parse_config")
-    @mock.patch("aprsd.log.setup_logging")
+    @mock.patch("aprsd.logging.log.setup_logging")
     def test_no_tocallsign(self, mock_logging, mock_parse_config):
         """Make sure we get an error if there is no tocallsign."""
 
@@ -76,7 +76,7 @@ class TestSendMessageCommand(unittest.TestCase):
         assert "Error: Missing argument 'TOCALLSIGN'" in result.output
 
     @mock.patch("aprsd.config.parse_config")
-    @mock.patch("aprsd.log.setup_logging")
+    @mock.patch("aprsd.logging.log.setup_logging")
     def test_no_command(self, mock_logging, mock_parse_config):
         """Make sure we get an error if there is no command."""
